Move the config structs and more code removal
