+++ lib/src/libp2p/connection/established.rs
@@ -226,58 +226,3 @@ pub struct Config<TNow> {
     /// Entropy used for the randomness specific to this connection.
     pub randomness_seed: [u8; 32],
 }
-
-/// Configuration for a request-response protocol.
-#[derive(Debug, Clone)]
-pub struct ConfigRequestResponse {
-    /// Name of the protocol transferred on the wire.
-    pub name: String,
-
-    /// Configuration related to sending out requests through this protocol.
-    ///
-    /// > **Note**: This is used even if `inbound_allowed` is `false` when performing outgoing
-    /// >           requests.
-    pub inbound_config: ConfigRequestResponseIn,
-
-    pub max_response_size: usize,
-
-    /// If true, incoming substreams are allowed to negotiate this protocol.
-    pub inbound_allowed: bool,
-}
-
-/// See [`ConfigRequestResponse::inbound_config`].
-#[derive(Debug, Clone)]
-pub enum ConfigRequestResponseIn {
-    /// Request must be completely empty, not even a length prefix.
-    Empty,
-    /// Request must contain a length prefix plus a potentially empty payload.
-    Payload {
-        /// Maximum allowed size for the payload in bytes.
-        max_size: usize,
-    },
-}
-
-impl ConfigRequestResponseIn {
-    /// Returns the maximum allowed size of a request.
-    ///
-    /// Returns `0` for [`ConfigRequestResponseIn::Empty`].
-    pub fn max_size(&self) -> usize {
-        match self {
-            ConfigRequestResponseIn::Empty => 0,
-            ConfigRequestResponseIn::Payload { max_size } => *max_size,
-        }
-    }
-}
-
-/// Configuration for a notifications protocol.
-#[derive(Debug, Clone)]
-pub struct ConfigNotifications {
-    /// Name of the protocol transferred on the wire.
-    pub name: String,
-
-    /// Maximum size, in bytes, of the handshake that can be received.
-    pub max_handshake_size: usize,
-
-    /// Maximum size, in bytes, of a notification that can be received.
-    pub max_notification_size: usize,
-}
